change delete

+++ src/routes/comment.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const Comment = require("../../database/commentdb");
+const CommentModel = require("../models/comment")
 const router = express.Router();
 
 router.post("/create", (req, res) => {
@@ -10,7 +11,7 @@ router.post("/create", (req, res) => {
         req.body.name,
         req.body.content,
         (result) => {
-            console.log(result);
+            // console.log(result);
             res.status(200).send()
         }
     );
@@ -21,19 +22,32 @@ router.post("/", (req, res) => {
     Comment.getComment(
         req.body.tableId,
         (item) => {
-        console.log(item)
+        // console.log(item)
         res.status(200).json(item)
     });
 });
 
 router.delete("/delete", (req, res) => {
     console.log("/comment/delete")
-    Comment.deleteOne(
-        req.body.name,
-        req.body.commentId,
-        (item) => {
-        console.log(item)
-        res.status(200).json(item)
+    CommentModel.find({"name": req.body.name, "_id": req.body.commentId}, (err, result) => {
+        console.log(result)
+        if (result.length == 0) {
+            console.log('A');
+            res.status(400).send()
+        } else {
+            console.log('B');
+            Comment.deleteOne(
+                req.body.name,
+                req.body.commentId,
+                (item) => {
+                // console.log(item)
+                if (item == null) {
+                    res.status(400).send()
+                } else {
+                    res.status(200).json(item)
+                }
+            });
+        }
     });
 });
