+++ src/routes/table.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const Table = require("../../database/tabledb");
+const TableModel = require("../models/table");
 const router = express.Router();
 
 router.post("/create", (req, res) => {
@@ -11,7 +12,7 @@ router.post("/create", (req, res) => {
         req.body.email,
         req.body.clubName,
         (result) => {
-            console.log(result);
+            // console.log(result);
             res.status(200).send()
         }
     );
@@ -20,7 +21,7 @@ router.post("/create", (req, res) => {
 router.get("/", (req, res) => {
     console.log("/table/")
     Table.getAll((item) => {
-        console.log(item)
+        // console.log(item)
         res.status(200).json(item)
     });
 });
@@ -30,7 +31,7 @@ router.get("/:tableId", (req, res) => {
     Table.getOne(
         req.params.tableId,
         (item) => {
-        console.log(item)
+        // console.log(item)
         res.status(200).json(item)
     });
 });
@@ -38,12 +39,23 @@ router.get("/:tableId", (req, res) => {
 router.delete("/delete", (req, res) => {
     console.log("/table/delete")
     console.log(req.body)
-    Table.deleteOne(
-        req.body.name,
-        req.body.tableId,
-        (item) => {
-        console.log(item)
-        res.status(200).json(item)
+
+    TableModel.find({"name": req.body.name, "_id": req.body.tableId}, (err, result) => {
+        console.log(result)
+        if (result.length == 0) {
+            console.log('A');
+            res.status(400).send()
+        } else {
+            Table.deleteOne(
+                req.body.name,
+                req.body.tableId,
+                (item) => {
+                console.log('B');
+                console.log(item)
+                res.status(200).json(item)
+                }
+            );
+        }
     });
 });
